Add support for VLANs (#145)
This PR adds support for VLAN tagging of overlay provider networks.

Overlay provider network VLAN tags are attached by OVN to all egress traffic from the VMs. This is accomplished by configuring the tag to the logical switch port of the external router port that connects from the project network to the provider network.

Bridge VLANs and IP mode
The bridge adapter is not affected by the provider's VLAN setting. The default configuration for an overlay adapter also doesn't enable the bridge adapter for network communications.

For scenarios where we have few (or only one) network adapters, the bridge can now be configured to be enabled for DHCP by default. For this case, we now also have the option to configure a VLAN for the bridge adapter:

* added provider vlan tagging
new field provider_vlan to tag traffic from a network provider
to a vlan.

* added bridge adapter vlan configuration
replaced enabled_adapters with a bridge option for
disabled bridges or dhcp enabled bridges.

* update .net OVN to rc.2