double check for Hyper-V
This PR adds a additional check for a working Hyper-V environment after feature installation. 

This prevents that eryph installation is started for two cases:
- Hyper-V feature is installed and script is started a second time before reboot
- Hyper-V feature is already installed but not functional - e.g. VMMS service is broken or disabled. 

Also reboot behavoir for feature installation was improved, as in desktop the script may stop asking for a reboot. This can only be controlled by disabling automatic reboot at all.