double check for Hyper-V (#301)
This PR adds an additional check for a working Hyper-V environment after the feature installation.

This prevents the eryph installation from starting in two cases:

Hyper-V feature is installed and script is started a second time before reboot.
Hyper-V feature is already installed but not functional - e.g. VMMS service is broken or disabled.
The reboot behavior of the feature installation has also been improved, as the script might stop asking for a reboot on the desktop. The only way to control this is to disable automatic reboot altogether.